+++ quickwit-actors/Cargo.toml
@@ -21,7 +21,6 @@ flume = "0.10"
 futures = "0.3"
 tracing = "0.1.29"
 quickwit-common = { version = "0.3.0", path = "../quickwit-common" }
-once_cell = "1"
 
 [dev-dependencies]
 rand = "0.8"
