+++ quickwit-common/src/lib.rs
@@ -19,10 +19,12 @@
 
 mod checklist;
 mod coolid;
+
 pub mod fs;
 pub mod metrics;
 pub mod net;
 pub mod rand;
+pub mod runtimes;
 pub mod uri;
 
 use std::fmt::Debug;
