+++ quickwit-indexing/src/actors/packager.rs
@@ -26,13 +26,13 @@ use anyhow::{bail, Context};
 use async_trait::async_trait;
 use fail::fail_point;
 use itertools::Itertools;
-use quickwit_actors::{
-    Actor, ActorContext, ActorExitStatus, Handler, Mailbox, QueueCapacity, RuntimeType,
-};
+use quickwit_actors::{Actor, ActorContext, ActorExitStatus, Handler, Mailbox, QueueCapacity};
+use quickwit_common::runtimes::RuntimeType;
 use quickwit_directories::write_hotcache;
 use quickwit_doc_mapper::tag_pruning::append_to_tag_set;
 use tantivy::schema::FieldType;
 use tantivy::{InvertedIndexReader, ReloadPolicy, SegmentId, SegmentMeta};
+use tokio::runtime::Handle;
 use tracing::{debug, info, info_span, warn, Span};
 
 /// Maximum distinct values allowed for a tag field within a split.
@@ -109,8 +109,8 @@ impl Actor for Packager {
         self.actor_name.to_string()
     }
 
-    fn runtime_handle(&self) -> RuntimeType {
-        RuntimeType::Blocking
+    fn runtime_handle(&self) -> Handle {
+        RuntimeType::Blocking.get_runtime_handle()
     }
 }
